move code away from main.go
